requirements.txt
certifi==2020.6.20
cycler==0.10.0
easydict==1.9
future==0.18.2
joblib==0.16.0
kiwisolver==1.2.0
lap==0.4.0
matplotlib==3.3.2
memory-profiler==0.57.0
numpy==1.19.2
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.13.0
psutil==5.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
xlwt==1.3.0
